+++ lib/src/cast/index.dart
@@ -7,3 +7,4 @@ export 'widgets/cast_button.dart';
 export 'widgets/cast_devices_dialog.dart';
 export 'widgets/cast_overlay.dart';
 export 'widgets/cupertino_cast_button.dart';
+export 'widgets/cupertino_cast_devices_sheet.dart';